VC++ again

+++ testing/adios2/engine/staging-common/TestCommonClient.cpp
@@ -257,7 +257,7 @@ TEST_F(SstReadTest, ADIOS2SstRead)
 
         if (myStart + myLength > writerSize * Nx)
         {
-            myLength = (long unsigned int)writerSize * Nx - myStart;
+            myLength = (long unsigned int)writerSize * (int)Nx - myStart;
         }
         const adios2::Dims start{myStart};
         const adios2::Dims count{myLength};
